+++ cloudauth-20190307/alibabacloud_cloudauth20190307/models/_id_3meta_verify_advance_request.py
@@ -11,6 +11,7 @@ class Id3MetaVerifyAdvanceRequest(DaraModel):
         self,
         crop: str = None,
         face_file_object: BinaryIO = None,
+        face_picture: str = None,
         face_url: str = None,
         identify_num: str = None,
         param_type: str = None,
@@ -27,6 +28,7 @@ class Id3MetaVerifyAdvanceRequest(DaraModel):
         # Portrait side of the ID card image input stream.
         # Choose one between `CertUrl` and `CertFile`.
         self.face_file_object = face_file_object
+        self.face_picture = face_picture
         # Portrait side of the ID card image.
         # Accessible HTTP or HTTPS link on the public network.
         # Choose one between `CertUrl` and `CertFile`.
@@ -64,6 +66,9 @@ class Id3MetaVerifyAdvanceRequest(DaraModel):
         if self.face_file_object is not None:
             result['FaceFile'] = self.face_file_object
 
+        if self.face_picture is not None:
+            result['FacePicture'] = self.face_picture
+
         if self.face_url is not None:
             result['FaceUrl'] = self.face_url
 
@@ -85,6 +90,9 @@ class Id3MetaVerifyAdvanceRequest(DaraModel):
 
         if m.get('FaceFile') is not None:
             self.face_file_object = m.get('FaceFile')
+
+        if m.get('FacePicture') is not None:
+            self.face_picture = m.get('FacePicture')
 
         if m.get('FaceUrl') is not None:
             self.face_url = m.get('FaceUrl')