+++ cloudauth-20190307/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-06-08 Version: 3.13.3
+- Update API Id3MetaVerify: add request parameters FacePicture.
+- Update API InitFaceVerify: add request parameters EnableBeauty.
+
+
 2026-04-14 Version: 3.13.2
 - Update API MobileRecycledMetaVerify: add response parameters Body.ResultObject.IspName.